- Eliminate non-portable snprintf(). Fixes Win32 MSVC build. - Remove some debug leftovers
SVN r1562 (odameXbox)

+++ client/sdl/i_input.cpp
@@ -248,7 +248,7 @@ CVAR_FUNC_IMPL (use_joystick)
 
 CVAR_FUNC_IMPL (joy_active)
 {
-	if( (var < 0) || (var > I_GetJoystickCount()) )
+	if( (var < 0.0) || ((int)var > I_GetJoystickCount()) )
 		var = 0.0;
 
 	I_CloseJoystick();
